feat: Add placeId and placeName properties to Address model
This commit adds the `placeId` and `placeName` properties to the `Address` model in the `address.dart` file. These properties will be used to store the unique identifier and name of a place, providing additional information about the address. This change enhances the functionality of the `Address` model and allows for more comprehensive address representation.