feat: Add service_account_key.json to .gitignore and pubspec.yaml
This commit adds the `service_account_key.json` file to the `.gitignore` and `pubspec.yaml` files. The `service_account_key.json` file contains sensitive information related to Firebase options and should not be tracked in version control. This update ensures that the file is properly ignored and not included in future commits.