feat: Update dependencies in pubspec.lock
This commit updates the dependencies in the pubspec.lock file. It includes updates to the versions of various packages such as flutterfire_internals, cloud_firestore, firebase_auth, and geolocator. These updates ensure that the project is using the latest stable versions of these packages, which may include bug fixes, performance improvements, and new features.