chore: Update Gradle and Kotlin versions
This commit updates the Gradle and Kotlin versions in the build.gradle file. It updates the Kotlin version to '2.0.10' and adds the Kotlin Gradle plugin as a dependency. Additionally, it updates the Gradle distribution URL in the gradle-wrapper.properties file to 'https://services.gradle.org/distributions/gradle-8.3-rc-2-bin.zip'. These updates ensure that the project is using the latest stable versions of Gradle and Kotlin, which may include bug fixes, performance improvements, and new features.